修复tcc编译失败的bug

+++ mkrtos_user/lib/libc_backend/inc/fd_map.h
@@ -3,8 +3,8 @@
 
 typedef struct fd_map_entry
 {
-    uint16_t svr_fd;
-    uint16_t priv_fd;
+    uint32_t svr_fd;
+    uint32_t priv_fd;
     uint8_t type;
     uint8_t flags;
 } fd_map_entry_t;
@@ -15,7 +15,7 @@ enum fd_type
     FD_FS,
 };
 
-int fd_map_alloc(uint16_t svr_fd, uint16_t priv_fd, enum fd_type type);
+int fd_map_alloc(uint32_t svr_fd, uint32_t priv_fd, enum fd_type type);
 int fd_map_update(int fd, fd_map_entry_t *new_entry);
 int fd_map_free(int fd, fd_map_entry_t *ret_entry);
 int fd_map_get(int fd, fd_map_entry_t *new_entry);
